+++ eng/scripts/check-for-changed-files.js
@@ -1,19 +1,6 @@
-import { run } from "./helpers.js";
+import { checkForChangedFiles, coreRepoRoot, repoRoot } from "./helpers.js";
 
-const proc = run("git", ["status", "--porcelain"], {
-  encoding: "utf-8",
-  stdio: [null, "pipe", "pipe"],
-});
-
-if (proc.stdout) {
-  console.log(proc.stdout);
-}
-
-if (proc.stderr) {
-  console.error(proc.stderr);
-}
-
-if (proc.stdout || proc.stderr) {
+if (checkForChangedFiles(coreRepoRoot, "## cadl ##", repoRoot, "## cadl-azure ##")) {
   console.error(
     `ERROR: Files above were changed during PR validation, but not included in the PR.
 Include any automated changes such as sample output, spec.html, and ThirdPartyNotices.txt in your PR.`